Exploring auto-commit and auto-offset-reset
Both auto-commit and auto-offset-reset are now CLI options.

Producers and consumers log their compact configs (just changes to the
defaults) once per run.

Fixes a bug in the queue workload which caused it to incorrectly tag
operations as :fail when some of their mops could have succeeded.

Fixes a bug in the checker which couldn't handle it when we read values
with no offset, and also couldn't infer the offset from a write.