Queue: catch (eventual) txn atomicity violations
We now use a mutable structure to track incremental completions of
micro-ops during txn, poll, and send operations, which means that even
when we crash, we return offsets/polls. This gives us the ability to
identify lost updates from indeterminate transactions, if we can prove
that other parts of that transaction were visible to other polls.